+++ .isort.cfg
@@ -7,7 +7,7 @@ import_heading_thirdparty=Third-Party Libraries
 import_heading_firstparty=cisagov Libraries
 
 # Should be auto-populated by seed-isort-config hook
-known_third_party=apps,aws_lambda_wsgi,boto3,dotenv,eal,marshmallow,motor,setuptools
+known_third_party=apps,aws_lambda_wsgi,boto3,dotenv,eal,marshmallow,setuptools
 # These must be manually set to correctly separate them from third party libraries
 known_first_party=api,authentication,bson,database,flask,faker,jwt,lmotor,names,pytest,pymongo,requests,schematics,service,simpleeval,test_validators,yaml
 # This helps black and isort get along
