+++ requirements.txt
@@ -45,7 +45,6 @@ requests==2.24.0
 ruamel.yaml==0.16.10
 ruamel.yaml.clib==0.2.0
 s3transfer==0.3.3
-schematics==2.1.0
 six==1.15.0
 toml==0.10.1
 traitlets==4.3.3
