+++ src/api/__init__.py
@@ -1,0 +1,3 @@
+"""initialize api app."""
+# cisagov Libraries
+from api.views import api
