--- src/apps/api/__init__.py
@@ -1,3 +1,0 @@
-"""initialize api app."""
-# Third-Party Libraries
-from apps.api.views import api
