--- src/apps/database/repository/__init__.py
@@ -1,1 +1,0 @@
-"""This is a blank init python file."""
